functions/event-trigger/requirements.txt
google-cloud-workflows